Move translating facets to separate service
